Fix bugs in initialization of _support_seccomp_user_notif and SECCOMP_FILTER_FLAG_NEW_LISTENER flags.
The _support_seccomp_user_notif should be initialized in
sys_chk_seccomp_action() rather than sys_chk_seccomp_flag() (the
latter is a flag check, not an action check).

When loading a filter, the seccomp SECCOMP_FILTER_FLAG_NEW_LISTENER flag
should only be set if the filter is using the notify action.

Note that filters that use the notify action must not use the tsync
flag (the user is required to disable tsync on such filters
explicitly via seccomp_attr_set()).